first implementation of SDL timers

+++ SDL_buttons_timers/manager_utils/inc/manager_utils/config/DrawMgrConfig.hpp
@@ -13,7 +13,7 @@
 namespace DrawMgrConfig {
 struct Config_t {
 	MainWindow::Config_t m_WindowCfg;
-	std::int64_t m_MaxFrameRate;
+	std::int32_t m_MaxFrameRate;
 };
 }
